sqlite updates
