+++ connection/connection.go
@@ -1,7 +1,7 @@
 /*
 Copyright © 2024 Zac Orndorff zac@orndorff.dev
 */
-package cmd
+package connection
 
 import (
 	"fmt"
